requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
motor==3.3.2
pymongo==4.6.1
aiohttp==3.9.1
aiofiles==23.2.1
Pillow==10.2.0
python-dotenv==1.0.0
uvloop==0.19.0
